BUG#31378900 - RESOURCE_GROUP_MGR HEAP-USE-AFTER-FREE.
MTR tests when run under no-threads connection handler fail with
ASAN build. This is becasue resource group is deinitialized for
non per thread connection handler. PFS thread create callback is
called and we access the freed Resource_group object to retrieve
the names of the default resource group to associate with the thread.
The fix is to make the default resource group names always available
to associate with resource group field of the PFS threads table.
This patch is based on the contribution of Xiaoyu Wang, Tencent Technology.

Change-Id: I1971b0572f3cae06ac06f1cc71813a75f5ecd8b0