Bug #31437753: HEAP-BUFFER-OVERFLOW IN HASH JOIN STUFF
If the sort key for a string hash join key can be over 1024 bytes, we stop
storing it in the hash table and ask for a (collation-aware) hash of the string
instead; this makes sure that we don't try to allocate multi-gigabyte sort keys
(it's very inefficient, and can also make for edge-case bugs). This mostly happens
with the old PAD SPACE collations. Include strings that are part of row keys in
this logic, for exactly the same reasons.

Change-Id: I457289021941b72b3ebdd88ed7f10b18f5d4aee9