Bug #31473945: CREATE --SKIP_IF_HYPERGRAPH MTR DIRECTIVE
Add an MTR directive called --skip_if_hypergraph, which has the same effect
as the hypergraph optimizer rejecting a query. The intended use is for a
statement that the optimizer won't reject by itself, but for whatever reason
will return a different result for. A typical example is handler call counts,
where the differing query plan would cause different values; skipping this
allows the test to pass with a skipped message instead of failing outright.
Another one is EXPLAIN, where the plan may be different (if nothing else,
in costs); TRADITIONAL and JSON are already auto-rejected, but TREE is not.

Note that this results in a test being marked as skipped (assuming everything
eles is OK), not passing. Many of these tests need to either be updated,
or the hypergraph optimizer be fixed to include the functionality tested by the
test. This is akin to the regular error rejects.

Demonstrate its use a number of places that are relevant for the hypergraph
optimizer.

Change-Id: I4e4a7ad5080b5b25356bb4901c618727ec181260