Bug#31406724: Crash in sql_scalar_to_json after many recursive calls
Based on a suggestion from Knut Anders Hatlen.

Evaluating a JSON scalar value may enter an infinite loop:
In sql_scalar_to_json(), when field type is MYSQL_TYPE_NULL,
the null value of the item is calculated, which will reach the same
function in an infinite loop. This was never supposed to happen,
however a parameter that is NULL now has the "actual type" set to
MYSQL_TYPE_NULL and null_value set to true.

The fix is thus to avoid calling arg->update_null_value() and
trusting that null_value has already been set.

Reviewed by: Catalin Besleaga <catalin.besleaga@oracle.com>