Bug#31469354 CLUSTERJ TEST FAILS IN SOLARIS 32 BIT VERSIONS IN PB2
The test fails due to some inconsistencies between the Java
installations in the 32 bit Solaris platforms and the MTR expectations.
This patch adds a check to skip running the tests in 32 bit platforms
through MTR for MySQL Cluster Versions 7.3 and 7.4. They can still be
run through Maven. This patch also backports parts of fix for
Bug#28424366 to change the minimum required Java version needed for
running the MTR ClusterJ tests to Java 1.8.

Change-Id: I17188a4008db9d36326dfecd8db4221232bbb253