Bug #31337487   NDB : LCP MAXDISKWRITESPEED BOUNDS NOT ACCOUNTING FOR OWN RESTAR
Prior to 7.6, there was no dynamism in the max disk write speed value
until a node had started (phase 7), so that it used
MaxDiskWriteSpeedOwnRestart for all LCPs during startup.

In 7.6, dynamism during restart was enabled, but this can result in
a lower disk write speed being chosen during restart.

This is fixed so that :
- The default 'max' during restart is MaxDiskWriteSpeedOwnRestart
- If some other mechanism decides that more speed is required,this is possible

Approved by : Mikael Ronstrom <mikael.ronstrom@oracle.com>