WL#9384 Prepare statement once
Problem with "read before write removal" for UPDATE with parameter(s) in
value list. The "read" was not removed since parameters was not properly
determined as constant, this caused additional roundtrips to storage
engine.

Fix by changing check_constant_expressions() to use the new
Item::const_for_execution() function when determining if the update
value list contains a constant expression.

Add tests both for constant parameters in update value list as well as
where clause. Remove some redundant eval and printout of variables which
won't be expanded in "die".

Change-Id: I6a042da1194c7f81d83723d7894635e9a50f5b8e