Bug#31481225 MOVE CHECK_IMPLICIT_COLUMN_FORMAT_CHANGE TO LOCAL SCOPE
The function check_implicit_column_format_change() is defined as a member
of the ha_ndbcluster class although it does not need any part of that
class, this increases the interface of ha_ndbcluster in an undesirable
way.

Change the check_implicit_column_format_change() function to a local
scope function defined in ha_ndbcluster.cc where the function is used.
Mark variables in the function which are assigned but never modified as
const.
Make the NDB_VERSION_DYNAMIC_IS_DEFAULT variable const.
Fix spelling errors in comments.

Change-Id: I032ee388b2d1be034b49419a3f89ce54356ae2eb