Bug#31356801 NO NEED TO INVALIDATE PARENT TABLE FROM TDC AFTER FK ADD/DROP
Before the introduction of MySQL Data Dictionary, ndbcluster used to
maintain a cache of the foreign key metadata of a table in its handler
object. This was provided to the server whenever it asked for any
information about the foreign keys. After any foreign key creation or
deletion, the parent table in that relationship was flushed out from the
TDC so as to ensure that the next time it is accessed, the updated
foreign key information is loaded into the handler along with the table.

But after MySQL DD, the server stopped asking for any foreign key
information from the SEs and the foreign key metadata cache stored in
ndbcluster's handler was removed as it didn't have a purpose anymore.
So, any invalidation of parent tables from TDC is also not required
anymore as there is nothing new to be loaded into the handlers. This
patch removes all such parent table invalidations.

Change-Id: Icab2b740f4ed2998a1eee870eced26c045e14262