Bug#31484875 PRINTOUT OF POINTER WHICH MAY BE NULL
In the function ha_ndbcluster::ndb_optimize_table() there is a loop to
optimize unique indexes of a table. Inside the optimize loop, there is a
printout which refers to a pointer variable from previous loop over
indexes, that pointer variable may be null if the table does not have
any such index. This will cause crash to occur in debug compiled
binaries.

Change the usage of "index" variable to "unique_index".
Move variables into scope as much as possible.

Change-Id: I3d6f93fd54c9bfaaf6f788d883d8ef0ee971384d