Bug#31437158	ASSERTION `JOIN == NULLPTR' FAILED AT SELECT_LEX::OPTIMIZE AT SQL_SELECT.CC
We execute a query.
First we prepare/optimize it for InnoDB.
Then we do the same against RAPID: MySQL chooses to do a hash join,
which tests if the join condition:
table1 . pk < ANY ( SELECT 3 UNION  SELECT 8 )
is always true; ConditionIsAlwaysTrue() is called for that, it evaluates
the UNION subquery. This subquery is thus optimized, and during this, when
optimizing fake_select_lex, RAPID says it's not supported.
So, SELECT_LEX_UNIT::optimize(), which gets a failure in
fake_select_lex->optimize(), returns with error before doing
'set_optimized()' on itself. This non-zero return value is, as usual,
"eaten" by the val_bool() function.
So the creation of iterators continues, and finally we get to the end of
the top query's SELECT_LEX's optimization.
Then we continue with optimizing its non-optimized subqueries, and
the UNION is seen as non-optimized, so we optimize it again and
fail an assertion as select_lex->join is already non-null.
Fix: after creating iterators, check thd->is_error() and make optimization
fail if it's true.
With this fix, the query will be re-routed to InnoDB in a third phase.

Change-Id: I3e0f0f21712d4817e38adc4066c406d07f0a4aa5