Bug #31491799: DYNAMIC_STRING HAS INEFFICIENT REALLOCATION
MySQL has its own string class called DYNAMIC_STRING, which is loosely like a
std::string but less-featured. When it needs to grow, it increases linearly by
a predefined amount of bytes (typically 128, occasionally as much as 4096) and
copies all the bytes over. This can lead to severe inefficiencies with repeated
appends over large strings, for instance in mysqltest.

Use exponential growth instead (1.5x each time we need to grow), like in STL.
We cannot easily replace all DYNAMIC_STRING uses with std::string, even though
it would be nice, as several of these objects are declared in static scope
(so having something with nontrivial destructors would be against the style guide).

Change-Id: If14d60bc4588a9276e4d8e53cf0a3fb09eff7673