Bug#31450875 ROUTERTEST_COMPONENT_METADATA_TTL FAILS ON SLOW MACHINES
The test_metadata_ttl assumes that a given number of refreshes
happen between the router is started and until the test stops
the Router. That also assumes that the refresh itself is
very quick and it's duration can be ignored. That does not hold
for the slower pb2 platforms where the refresh procedure itself
can take up to hundreds of milliseconds.

The tests were refactored to use the log timestamp to determine
how long the Router waits between consecutive metadata refreshes
instead of counting them over the timespan and assuming how many
there should be.

RB: 24600