WL#14142 UPDATE CMakeLists.txt FOR -devbld BUILDS
Updates CMakeLists.txt of legacy versions(7,3 and 7.4) of cluster for
-devbld builds to build NDB tests and NDBAPI examples to increase test
coverage.

-devbld builds were requested from PB2 so that Autotest could use the
same binaries for running Autotest tests. But, test binaries were not
being generated even when -devbld was enabled because it wasn't handled
in the CMake files.

So, CMakeLists.txt is modified to build with WITH_NDB_TEST=1 and
WITH_NDBAPI_EXAMPLES=1 when built with -devbld suffix.

Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>

Change-Id: Iaf763a6adbf916a177cafb996e11765c56097c5f