BUG #31468590 ASSERTION FAILED: FILE->M_FILE_STAT.M_OPEN_COUNT <= 1
Description
-----------
On Windows, it is possible for the GetTempFileName function to
reuse a previous filename if it is called twice in quick succession
with the first file being deleted before the second call.
MySQL doesn't expect this filename reuse behaviour, hence the
assertion failure.

Fix
---
This patch modifies Windows temporary filename creation to use
UuidCreateSequential for part of the temporary filename, relying
on this function's stronger guarantees of uniqueness to avoid
filename reuse.

As the temporary filenames are now slightly longer than previously,
this patch also adds a new check on Windows of temporary directory
lengths at MySQL server startup, aborting startup with
ER_TMPDIR_PATH_TOO_LONG if necessary.

RB: 24676