BUG#31446381: ERROR WHEN PROCESSING CERTIFICATION INFORMATION IN THE INCREMENTAL RECOVERY PROC
When a member joins a Group Replication group, it goes through the
distributed recovery procedure, which is composed by a optional
clone step and a final incremental step.
On the clone step, the member GTID_EXECUTED is only updated after
the server restart that finishes the clone step.
During the clone step the member will periodically send the
GTID_EXECUTED value that the server had once joined the group.
This will restrain the common set of transactions applied on all
members, which in consequence will render the certification garbage
collection void.
If the clone step takes a considerable amount of time, the
certification info, which is not being garbage collected, may
increase to a point on which will be impossible to send it to the
member after the server restart.

To avoid the above issue, now we only consider ONLINE members for
the common set of transactions applied on all members.
When recovering members change to ONLINE state, their certification
info will be updated with the one of the donor at the join, being
garbage collect on the future garbage collection rounds.

RB: 24656