Bug#31357022 FIX REPLICATION TESTS FOR WHEN RUN WITH --REPEAT=N (N>1)
[Post fix]

Due to some incompatibilities with NDB, the cleanup procedure
implemented on the base patch caused some issues on NDB tests where
asynchronous replication is used.

This issue is related to RESET SLAVE ALL described on Bug#31515760
SLAVE RESTARTS ARE NOT PROPERLY DETECTED BY NDB AFTER A RESET SLAVE
ALL

To fix this problem for now, we revert the RESET SLAVE behavior but
only for NDB setups.

ReviewBoard: 24655