Bug #30963751 TEST CASE INNODB.ZLOB_IBD_SIZE IS FAILING SPORADICALLY IN PB2
Problem:

The description of this bug provides a call stack which points to a problem
in rollback when LOB is involved.  But this is not seen anymore.  But the
sporadic failure of test innodb.zlob_ibd_size is seen for a different reason
- timeout in the include/wait_innodb_all_purged.inc script.

Solution:

Currently we give 60 seconds for the script include/wait_innodb_all_purged.inc
to complete. It is simply possible that this time is not sufficient when it
is a slow machine.  Also, tests that run prior to innodb.zlob_ibd_size could
leave more data to be purged.

So, tentatively I am doing the following to solve this issue:

1.  Increase the timeout from 60 to 100 seconds.
2.  Introduce a new call to include/wait_innodb_all_purged.inc at the
    beginning of this test to clear the data left from previous mtr tests.

rb#24657 approved by Kevin Lewis <kevin.lewis@oracle.com>