BUG#31505890: USAGE OF IS_EXECUTED BY GROUP REPLICATION LACKS LOCKING
Group Replication consistent reads feature needs to check if a given
transaction id (Gtid) is already committed. For that it uses the
method:
  bool Gtid_state::is_executed(const Gtid &gtid)
which requires that the `global_sid_lock` and `sidno` lock of
the queried Gtid are acquired. Though the `sidno` lock was not being
acquired leaving this query vulnerable to concurrent updates on the
GTID_EXECUTED set.

To fix the above issue, now we do acquire both locks.

RB: 24669