WL#14141: Make a JSON logsink suitable for OCI logging
This WL adds a new error-logging component that writes a
specialized log for OCI.

The component gets set up automatically in its new home.
If you need to test it locally, kindly see the enclosed
mtr test case and the WL ticket on how to set up and
use the component (standard component procedure applies,
but in addition certain variables need to be set before
logging can commence).

Change-Id: I337a9396e1cfbd3cf04c4027c0f2fda9914f3007