Bug#30950714 CONSISTENT SNAPSHOT CAN BE CORRUPTED BY OTHER TRANSACTIONS
Issue
=====
We make use of prebuilt->m_end_range flag to check if end_range condition is
not satisfied if the number of uncommitted rows we've passed during the scan
is > 100 to avoid full table traversal. In the call to row_search_mvcc if we
see that the flag is set we unset it and immediately return indicating that
there are no more row to be fetched for the the given range.

This is benefecial if we have prefetched rows that we need to return to the
server. In such a case, every time the server calls row_search_mvcc, after
fetching the first row, if there are any cached rows we return the cached
row, and when we've returned all the cached rows we check if the end_range
flag is set and if it's we unset it and return.

Turns out that this flag is not reset at the beginning of every index read
resulting in some reads being aborted because of this and resulting in
missing rows. One such case is when optimiser decides to use
Multi-Range-Read (MRR) optimisation as the range query specified by the
user is split into multiple range queries.

Fix
===
In row_search_mvv() reset the prebuilt->m_end_range cache at the start of
every index read.

RB: 24201
Reviewed-by: Debarun Bannerjee <debarun.bannerjee@oracle.com>