Bug #31310238: SUPPORT SORTING MULTIPLE TABLES [patch 5/5, outer joins with addons]
Similarly to the last fix, we can have outer joins with addons instead
of row ID, so we need to store the row ID flags for those rows, too.
This is required so that we can store NULLs for non-nullable columns;
in the case of packed addons, it also saves space.

Unlike the row ID case, we already have a good place to store individual
NULL bits, so we don't use a full byte for each table NULL row flag.
Note that this enables an edge case; we could have zero addon fields
but non-zero table NULL row flags, so we could invoke the addon packing
code without any addon fields. A lot of code assumed it could use the
offset of the first addon field (which may no longer exist) to find the
end of the NULL field zone, so that has been cleaned up. (Actually,
we don't need to store offsets for any field except the first,
so we can remove the field altogether.)

Similarly, we now model the length field for packed addons explicitly,
instead of going in and hacking the offsets directly.

Change-Id: I5b5e29d3a1d3290518bdd0f159b9455a4265ac51