Bug#31357022 FIX REPLICATION TESTS FOR WHEN RUN WITH --REPEAT=N (N>1)
[Post fix]

In a test that uses extra configurations, the group replication
applier was being reset on a scenario where a lack of disk space was
being simulated what leads to a failure.

This failure was not present before as we would reset the slave on
rpl_init.inc. To solve this issue for now we add a RESET SLAVE
instruction to the beginning of the test.

Approved by: Nuno Carvalho <nuno.carvalho@oracle.com>