Bug#31389573: STACK OVERFLOW IN MYSQLD
When we have windowing functions but no windowing steps, there is no iterator
to initialize the windowing function (the window is “shortcut”), and we must
iterate the Item in the top of a query block in order to avoid reading
uninitialized data, possibly causing crashes. This was done correctly only
for the top-level query block of a query expression, not for materialized
query blocks (such as derived tables).

Fix by copying the initialization code from
SELECT_LEX_UNIT::ClearForExecution() into
MaterializeIterator::MaterializeQueryBlock(). As the comment says, it would be
nice to do this in some iterator instead, but it's not immediately obvious
whether there would always be e.g. a FakeSingleRowIterator in such query blocks,
so this is the simplest and safest solution for now.

Change-Id: I63b3c9fa60002cc0171690d33add99fd71c6d8f2