Bug #31006095 - Assertion failure: TRX_SYS_ANY_ACTIVE_TRANSACTIONS() == 0
- Cleanup in certain replicated XA transactions would fail to reattach
InnoDB trx_t backup causing a leaked trx_t object which was detected on
server shutdown by the assertion in the trx sys.
- There were several failure modes, the unflag_engine_ha_data_detached
function was implemented with internal state, it would only return true
the first time it was called after a detach. In certain code paths, we
called it without calling re-attach, which obviously leads to a faulty
state.
- An additional problem was calling reattach_engine_ha_data_to_thd
for handlerton instances registered in the transaction context. This
method relies on the transaction context to get ha_info, while in xa.cc
there was a comment stating that even if we don't have ha_info in the
transaction context we still need to reattach the transaction using
thd->rpl_reattach_engine_ha_data() which sets the detached flag
internally.
- Fixed all these issues by removing state from the method that
checks whether the state is detached, and calling
rpl_reattach_engine_ha_data whenever we need to perform reattachment.

RB#24761 approved by Sven Sandberg<sven.sandberg@oracle.com>