Bug #31559245: REGRESSION: CRASH/ASSERT IN SUBSELECT_HASH_SJ_ENGINE::ROOT_ACCESS_PATH
If we know that a join will return zero rows, we don't bother
optimizing any IN expressions in the select list, but end optimization
early (see SELECT_LEX::optimize()). If so, don't attempt to print
them either, as they have no query plan.

This was a pre-existing issue even before WL #14061 (access paths),
but surfaced only then. The reason is that before, we would replace
the root iterator with a ZeroRowsIterator, but forget to copy over
join_for_explain(), so we wouldn't attempt to print iterators from
the SELECT list. After 14061, we have a more robust mechanism
that does not entail storing the JOIN pointer on iterators, so we
need to fix this more explicitly.

Change-Id: I79da15b16df865d3e6cb904799b2d7f07a3b8df3