Bug#30684839 SCHEMA DISTRIBUTION ABORT IS NOT HANDLED PROPERLY SOMETIMES
Problem :

During schema distribution, if the client gets killed after the DDL has
already been logged to the ndb_schema table, but before the participants
could reply, the client simply marks all the participants as failed in
the NDB_SCHEMA_OBJECT and returns. Since the distribution protocol was
already in progress, the coordinator still waits for the participants,
receives their ndb_schema_result insert and processes them. Meanwhile,
the client is now open to send in another DDL. If another DDL is
executed from the client and started to distribute, before the
coordinator can finish processing the previous schema change, the
DBUG_ASSERT() mentioned in the bug, which asserts that there is only one
schema op distribution active at a time, fails.

As the client returns, once it detects a thread kill, it also unlocks
the GSL. This might also lead to undefined issues as the participant is
still making the changes with an assumption that the GSL is being held
by the coordinator.

Solution :

Even if the thread has been killed, the client should not return after
the DDL has been logged to the ndb_schema table. Once the DDL has been
logged, coordinator is in control and the client simply waits for it to
make a decision.

Coordinator aborts the distribution only when it detects a server
shutdown or a cluster failure. Otherwise normally, it either waits for
all the participants to reply or to timeout to mark the schema op as
completed.

Change-Id: Iaa1756de0f1873992b0cff09e2c1331258d79192