Bug#31575027 data-race on g_reopen_thread
Problem
=======

g_reopen_thread is a global thread which watches for reopen requests for
the logfile. It waits for notifications from the SIGHUP signal handler.

The signal handler for SIGHUP is started before g_reopen_thread is
initialized.

TSAN reports a data-race on g_reopen_thread the signal-handler and
the initializing of g_reopen_thread.

Change
======

- add and check mutex around g_reopen_thread.

RB: 24735