WL#13686 Transform scalar subqueries to inner derived tables: with cardinality runtime check
This patch extends the capability of transforming scalar subqueries to
joins with a derived table. This is default for secondary engines like
RAPID, and can be enabled with an optimizer switch for primary engines
like InnoDB.

We now allow transformation of eligible scalar subqueries that are not
implicitly grouped as long as they are not correlated.  This is
implemented in primary engines as a run-time check that the derived
table join with has no more than one row.

Note that this WL does not add such run-time checks to RAPID. This
must be separately implemented to preserve SQL semantics.

For details see the WL description.

Reviewed-by: roy.lyseng@oracle.com

Change-Id: I350d8f4ed9ebd6531f88d74503e00e508bcfd670