Bug#31360425: Segfault in Query_result_update::optimize()
An update function used TABLE_LIST::updating instead of
TABLE_LIST::is_updated(), thus a table that was only selected from
was erroneously handled as an updated table.

The updating member differs from is_updated() in that it is also set
when FOR UPDATE is specified for a query block.

Fixed by replacing updating with is_updated() in sql_update.cc.

Reviewed by: Guilhem Bichot <guilhem.bichot@oracle.com>