WL#13860 Step 3: Node ID Allocation in mgmd.
Change the management server to use LocalDnsCache when resolving
hostnames after a node id allocation request, and to return
a new non-fatal error NDB_MGM_ALLOCID_CONFIG_RETRY in the case
where a hostname found in configuration could not be resolved.

This patch includes some rather significant refactoring in
match_hostname() and find_node_type(). The new code should be
easier to read, but verifying that the logic is equivalent in
all of the old cases, just by reading the patch, is a challenge.
Regression testing (and possibly unit testing) are important.

Note that there is no need to test AllowUnresolvedHostnames here;
if it were set to false and some names could not be resolved, the
server would have already shut down.

Reviewed-by: Sanjana DS <sanjana.ds@oracle.com>