bug#31343524 SPJ QUERIES HIT 'JOB BUFFER FULL' WHILE PROCESSING LQHKEYCONF
Remove existing solution for avoiding such 'job-buffer full' situations.
That solution has proven to be insufficient, see enhanced test case in previous
patch.

The existing solution was based on that we didn't start multiple lookup-sibling
operations in parallel, If such parallel lookup operation was needed, only KEYREQ's
to the first of these was sent when result from the ancestor scan node arrived.
The rest of the lookup operation nodes were deferred, and restarted when KEYCONF/REF
arrived from the previous completed operations.

That allowed a 'batchfull' of KEYREQs to be active simultaneously. However,
with large batchSizes, and the 'multiFragment-WL' allowing multiple fragments
to be scanned in parallel from the same SPJ, we could still overflow the job buffers.

Reviewed by: Mikael Ronstrom <mikael.ronstrom@oracle.com>