WL#9770 Keyring plugin for Oracle Public Cloud key store
Implement a MySQL keyring plugin talking to an Oracle Public Cloud's Infrastructure Vault back-end.

https://docs.cloud.oracle.com/iaas/Content/KeyManagement/Concepts/keyoverview.htm

MySQL provides keyring plugins for both the OKV and the competing AWS KMS. The OCI Vault is a natural fit for these secrets. So this plugin is a recommended addon for
MySQL OCI instances. It can work on premise too. And be a great gateway to OCI. Standalone hardware key servers can be expensive. If you trust OCI with your keys, then the
OCI can be a great substitute for the expensive on-premise hardware appliance.

RB:24444