bug#31343524 SPJ QUERIES HIT 'JOB BUFFER FULL' WHILE PROCESSING LQHKEYCONF
Prepare for a more dynamic decision of whether a particular row need
to be buffered or not.

New congestion control algorithm may decide on a row-by-row basis whether
a particular row need to be buffered for later resume of the operation.
Thus we may always need the rowBuffers for any TreeNode, and being prepared
to release rowBuffers.

Patch makes the relevant code unconditionally set up such buffers.

Reviewed by: Mikael Ronstrom <mikael.ronstrom@oracle.com>