bug#31343524 SPJ QUERIES HIT 'JOB BUFFER FULL' WHILE PROCESSING LQHKEYCONF
Patch introduce the new congestion control mechanism to avoid 'job buffer full'.

The old mechanism effectively limited the number of lookup-TreeNodes which could
be executed in parallel to '1'. However, we could still execute upto
'maxbatchSize' x 'max number of fragments' LQHKEYREQ's in parallel.
As demonstrated by new test case in the patch serie, that could be
sufficent to still overflow the job buffers. In addition it does not
make the SPJ a good block-thread-citizen to allow it having thousands of
outstanding signals in the job buffers.

This patch introduce SPJ congestion control based on the SPJ Requests
number of 'm_outstanding' signals instead. A 'HighlyCongestedLimit'
(256 signals) is defined: If number of outstanding signals in a Request
exceeds that, no further LQHKEYREQ's will be sent. Instead the
parent-row which the LQHKEYREQ is produced from will be buffered, and
the correlation-id of this row will be stored in the 'm_deferred' collection
of operations to later be resumed.

Resume of such deferred operations are triggered by ::lookup_countSignal(),
which are called whenever a LQHKEY_CONF/_REF and a TRANSID_AI from a lookup
TreeNode is received. The condition for more operations to be resumed is
that number of outstanding signals has decrease to 'MildlyCongestedLimit'
(Set to 50% of 'HighlyCongestedLimit'). This allowes for more LQHKEYREQ
to still being sent when we are in the MildlyCongestedLimit - HighlyCongestedRange,
without having the additional overhead of deferring the operation.

Patch has been checked against the TPC-H test suite without finding any
performance regressions. (Due to possible reduced concurrrency due to
less work being submitted in parallel).

Reviewed by: Mikael Ronstrom <mikael.ronstrom@oracle.com>