Bug #30175483: ASSERTION "N < M_SIZE" WITH SORT_BUFFER_SIZE AND MAX_SORT_LENGT SET
Description:
Assertion fialure for a query with ORDER By clause when server is
started with SORT_BUFFER_SIZE = 1125899906842624 and MAX_SORT_LENGTH
SET is set.

ANALYSIS:
If SORT_BUFFER_SIZE and MAX_SORT_LENGTH SET is set with some value
because of which variable Sort_param.max_keys_per_buffer in
filesort() can get value '0' then it asserts at statement
"n < m_size" in "Bounds_checked_array<unsigned char*>::operator[]".
So reason for this is we allow to create sort buffer of size '0' but
we assert when it's size is '0'.

SOLUTION:
Changed the following statement in filesort().

-     if (table_sort.get_sort_keys())
+      if (table_sort.sort_buffer_size() > 0 && table_sort.get_sort_keys())

This change will ensure we continue to search and allocate the sort
buffer if it's value is not greater than '0'.

Change-Id: I87d6a19839a7476d09aca49b69e6a3feefdd8bde