Bug#31587493:CONSTANT WHERE CONDITIONS ARE PUSHED DOWN TO DERIVED TABLE              EVEN WHEN NOT NEEDED
Problem:
While pushing conditions down to a derived table, currently any const condition
like a where "false" or a where "true" is pushed down to the first table in
the derived table. This does not manifest into a real problem. However this
condition pushdown is not necessary as the const condition has nothing to
do with the derived table.

Fix:
Avoid pushing const conditions down to derived table.
Additional fix:
We also update used tables for the condition the needs to be pushed down to
the derived table. Code inspection revealed that this condition does not
update used tables after replacing the columns in condition with the
derived table expressions.

Change-Id: I6cf947d449b18cecd08969bc5dd09d9a0dc3636f