Bug#31587575 CONDITIONS HAVING RAND() ARE NOT PUSHED DOWN TO DERIVED TABLE EVEN WHEN SAFE
Before this code change, the query at derived_condition_pushdown.test:313,
which has two AND-ed condition parts, had the RAND unpushed, and the
<10 pushed to WHERE.
After this code change, we see RAND is pushed to HAVING.
It is ok to push a non-deterministic condition down, but stopping
before the point where it would be calculated before the last
operation in the derived table. So, if there are WFs, we cannot push
it; if there is GROUP BY, we cannot push it to WHERE.

Regarding the change to queries using p(), it is because, with the
code change, p() became pushed to WHERE in both cases (when p() is
deterministic and when it's not: in the absence of WF and GROUP BY,
this now makes no difference), so I changed the query to use GROUP
BY, so we see the difference induced by determinism.

Change-Id: Id2960fec00ab2b0453bc4111d34d217cab23bbca