Bug#31597874 ROUTER CAN MISS REQUEST TO REFRESH METADATA.
There is a race condition that can cause the Router miss the request to
refresh metadata (for example triggered by the GR notification).
This is most likely to happen when the Router already is in the process
of the metadata refresh. The problematic part is the use of condition_variable
notify/wait. The thread that requests the refresh calls notify_one()
assuming that the refreshing thread is in the wait() and will unblock.
That does not have to be the case, the thread doing the refresh can be
doing something else at that point, like be in the middle of the metadata
refresh triggered earlier.

This patch adds a check if another metadata refresh was not requested
before blocking on cv.wait().

RB: 24780