Bug #31600044 ADD VERSION CHECK FOR SYSTEM ICU LIBRARY
Add explicit cmake version check for ICU library if configured with -DWITH_ICU=system.

This avoids build failure later, in particular on RedHat 7:
error: 'U_REGEX_INVALID_CAPTURE_GROUP_NAME' was not declared in this scope

The lowest version number checked, which also builds and runs OK,
is 55 (on Ubuntu 16), so set that as minimum required version.

A couple of pointer_cast<> rather than static_cast<> were needed to build with ICU 55.

Change-Id: I9585e47599e4641ddf98fa2e9178b42bbcff6243