Bug#31401637: Asan heap-buffer-overflow Handler::index_only_read_time
This problem may occur for a prepared statement that is executed
twice with different parameter sets, so that different plans are
generated for those executions.
The statement must also contain a materialized derived table for
which a generated key is made for the first execution.
What happens then is that the key information is not cleared from
the TABLE object representing the derived table.
In make_join_select(), we may then call test_quick_select(), which
may call find_shortest_key() since it thinks it has a key
(head->covering_keys.is_clear_all() is false). The function returns
MAX_KEY (as in no key found), however, there is no check that
the key is valid, and we continue, eventually indexing some struct
with MAX_KEY (=64) and we have the heap buffer overflow.

The problem is that the key information is not cleared after
execution. We thus add clearing of the key members of class TABLE
for materialized derived tables when destroying the JOIN object.

An assert is also added to check that the key returned from
find_shortest_key() is valid.

Reviewed by: Dag Wanvik <dag.wanvik@oracle.com>