Bug#31557370 - NDB.NDB_DD_SQL_FEATURES FAILS IN PB2
Inserting rows into a blob table results in "table is full" error,
showing lack of available extents for disk data.

The reason would be a delay in releasing the used extents by the
preceeding drop tables.

Table space with 1MB extent size will give 30(-1 reserved) extents
available, and would be easily consumed.

Since the table is inserted with a small #rows, reducing the extent
size to 256KB will give 120 extents available for inserts and may
compensate the not-yet-released extents.

The problem is noticed only in 7.3. The bug is fixed as described
above in later versions.

Fix: Integrate the above fix in 7.3 and 7.4. Add a debug to check
the #extents available in the test.

Reviewed by: Priyanka Sangam <priyanka.sangam@oracle.com>