WL#14166: xproto: notify on connection close
Motivation
==========
When a client sees a connection close from the server, it may have
several reasons: server shutdown, server crash, network loss,
connection idled too long, connection kill.
In case of a idling connection or killed connection, the server is still
running and the client just can reconnect to it.
In the other cases, a client may want to connect to other servers.

Goal
====
Allow an xprotocol client to distinguish:
- a recoverable connection loss
- a server failure.

Approved-by: Pavithra Pandith <pavithra.pandith@oracle.com>