Bug#31564407 TRX_T::ISOLATION_LEVEL IS NOT IN SYNC WITH THD->TX_ISOLATION
Issue
=====
As part of ha_innobase::sample_init() we set the InnoDB trx's isolation
level (trx_t::isolation_level) to READ UNCOMMITTED to avoid unnecessarily
looking up old versions of a record as the version list can be very long.

The issue is that this is not being reset after sampling which results in
assert being hit in ha_innobase::external_lock().

Fix
===
Once histogram sampling is done, reset the isolation_level of the trx.

RB: 24773
Reviewed-by: Annamalai Gurusami <annamalai.gurusami@oracle.com>
Reviewed-by: Jakub Lopuszanski <jakub.lopuszanski@oracle.com>