Bug #32112060 MYSQLROUTER CODE PARTIALLY SUPPORTS UNUSED METADATA VER 1 REPLICASETS
The metadata version 1 had a replicaset concept that was supposed to be used
to support multiple clusters.
This was never supported yet the data structures in the code were partially
prepared for that. This concept was removed in metadata version 2.
This patch simplifies the data structures in the code by removing the
replicaset concept.

RB: 26091