Bug #28445223  MTR'S "NO-SKIP" OPTION WORKS SILENTLY
Issue:
------
When --no-skip option prevents a test from being skipped, this decision happens
silently, with no indication in the output or error log.

Fix:
-----
In mysqltest client added exit code 63 to indicate "Attempted to skip, then passed"
and 64 to indicate "Attempted to skip, then failed ". Show this in MTR console log
as [ noskip-pass ] or [ noskip-fail ].

Reviewed-by: Deepa Dixit <deepa.dixit@oracle.com>
Reviewed-by: Srikanth BR <srikanth.b.r@oracle.com>