Bug #27533538 	INVALID THREADCONFIG ERROR
The data node fails startup with a crash when configured with
RealtimeScheduler=1. Index build during startup is done by
temporarily using some IO threads as index build threads. These
index build threads inherit the realtime properties of the IO
threads, so the index build threads are all realtime threads. When
the thread specification is checked, the index build threads are
checked to ensure that they are not realtime threads. The failure
of this check is treated as a fatal error and causes a crash.

This is fixed by always setting index build threads as non-realtime
threads instead of making index build threads inherit the realtime
settings of their host IO threads.

Approved by : Mikael Ronstrom <mikael.ronstrom@oracle.com>

Change-Id: I12852375e35aa97b32936cd9fa7b24ea338b9f3a