Bug #31588831: ASSERTION FAILURE IN NEWWEEDOUTACCESSPATHFORTABLES() AT SQL/SQL_EXECUTOR.CC
After Bug #31310238 (SUPPORT SORTING MULTIPLE TABLES), we would call
Filesort::using_addon_fields() fairly early, which decides whether
to use addon fields or not on the first call. This is problematic if
we sort a single table; we would need to know at that point whether
the table is under a weedout operation or not (because weedout forces
row IDs).

We already had logic in place to check this; however, ConnectJoins()
can add expand a weedout operation after-the-fact if there are non-hierarchical
weedouts present in the query. If so, we need to undo the addon field
decision, or we could be without row IDs for the table in question.

Change-Id: Iefd05afd768987565ec6a16b6e02287188b9e522