Bug#31470422 - CUSTOM FULLTEXT COLLATION CAUSES CLIENT AND SERVER SEGV IN 8.0.20
Cause:

While adding a collation, if we encounter an existing entry (collation
name) we skip further processing of the collation. This causes two
issues for the user defined collations that are loaded in two phases

1. In the first phase, during the server/client startup collations from
   the configuration file("Index.xml)" are loaded. If we encounter a
   duplicate of a inbuilt collation, the further processing of the
   collation is skipped. This leaves stale values("state") in the xml
   parser structure("cs"). These stale values cause uninitialized fields
   which force a server/client exit during shutdown.

2. in the second phase when we try to load the user collation map from
   the character set specific configuration file, we wrongly assume that
   it is overwrite of a inbuilt collation and skip loading the collation
   map. This makes the user collation unusable.

Fix:

In the "add collation" call (i) make sure "flags" is cleared every time
a collation in the configuration file is examined (ii) disallow
overwriting of compiled collations.

Change-Id: I84e6543c9ea43b8302b0ae975bdf43f8407cceff